# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgencyOS API
servers:
  - url: /api
...
paths:
...
  /quotations:
    get:
      operationId: listQuotations
      tags: [quotations]
      summary: List quotations
      responses:
        "200":
          description: Quotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quotation"
    post:
      operationId: createQuotation
      tags: [quotations]
      summary: Create quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
  /quotations/{id}:
    patch:
      operationId: updateQuotation
      tags: [quotations]
      summary: Update quotation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
    delete:
      operationId: deleteQuotation
      tags: [quotations]
      summary: Delete quotation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /quotations/{id}/convert-to-invoice:
    post:
      operationId: convertQuotationToInvoice
      tags: [quotations]
      summary: Convert quotation to invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchase-orders]
      summary: List purchase orders
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchase-orders]
      summary: Create purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchase-orders]
      summary: Get purchase order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    patch:
      operationId: updatePurchaseOrder
      tags: [purchase-orders]
      summary: Update purchase order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    delete:
      operationId: deletePurchaseOrder
      tags: [purchase-orders]
      summary: Delete purchase order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Invoice:
      type: object
      required: [id]
      properties:
        id:
          type: string
        number:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        companyGstin:
          type: ["string", "null"]
        clientGstin:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        termsAndConditions:
          type: ["string", "null"]
        bankDetails:
          type: ["object", "null"]
        lineItems:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/LineItem"
...
    LineItem:
      type: object
      properties:
        description:
          type: string
        hsnSac:
          type: ["string", "null"]
        qty:
          type: number
        unitPrice:
          type: number
        taxPercent:
          type: number
...
    Quotation:
      type: object
      required: [id]
      properties:
        id:
          type: string
        number:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        lineItems:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/LineItem"
        createdAt:
          type: string
    QuotationInput:
      type: object
      properties:
        clientId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        lineItems:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/LineItem"
...
    PurchaseOrder:
      type: object
      required: [id]
      properties:
        id:
          type: string
        number:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        orderDate:
          type: ["string", "null"]
        deliveryDate:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        vendorGstin:
          type: ["string", "null"]
        companyGstin:
          type: ["string", "null"]
        clientGstin:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        termsAndConditions:
          type: ["string", "null"]
        bankDetails:
          type: ["object", "null"]
        lineItems:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/LineItem"
        createdAt:
          type: string
    PurchaseOrderInput:
      type: object
      properties:
        clientId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        orderDate:
          type: ["string", "null"]
        deliveryDate:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        vendorGstin:
          type: ["string", "null"]
        companyGstin:
          type: ["string", "null"]
        clientGstin:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        termsAndConditions:
          type: ["string", "null"]
        bankDetails:
          type: ["object", "null"]
        lineItems:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/LineItem"
...